[dask] Refactor for learning to rank.
- Move data handling code into an independent module.
- Move ranking tests into an independent module.
- Use global pylint disable for import error.
- Remove old ranking demo.
