--- demo/rank/mq2008.conf
@@ -1,26 +1,0 @@
-# General Parameters, see comment for each definition
-
-# specify objective
-objective="rank:pairwise"
-
-# Tree Booster Parameters
-# step size shrinkage
-eta = 0.1
-# minimum loss reduction required to make a further partition
-gamma = 1.0
-# minimum sum of instance weight(hessian) needed in a child
-min_child_weight = 0.1
-# maximum depth of a tree
-max_depth = 6
-
-# Task parameters
-# the number of round to do boosting
-num_round = 4
-# 0 means do not save any model except the final round model
-save_period = 0
-# The path of training data
-data = "mq2008.train"
-# The path of validation data, used to monitor training process, here [test] sets name of the validation set
-eval[test] = "mq2008.vali"
-# The path of test data
-test:data = "mq2008.test"
