+++ python-package/pyproject.toml
@@ -21,6 +21,7 @@ classifiers = [
     "License :: OSI Approved :: Apache Software License",
     "Development Status :: 5 - Production/Stable",
     "Operating System :: OS Independent",
+    "Typing :: Typed",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.10",
@@ -59,10 +60,10 @@ follow_imports = "silent"
 ignore = ["tests"]
 extension-pkg-whitelist = ["numpy"]
 disable = [
+    "import-error",
     "attribute-defined-outside-init",
     "import-outside-toplevel",
     "too-many-nested-blocks",
-    "unexpected-special-method-signature",
     "unsubscriptable-object",
     "useless-object-inheritance"
 ]
